websocket_python/requirements.txt
requests
json
hmac
hashlib
logging
websocket-client
logging